+++ packages/form/src/Select/Select.js
@@ -5,7 +5,14 @@ import InputLabel from '../InputLabel';
 import Icon from '../../../icon/src/Icon';
 import Spinner from '../../../spinner/src/Spinner';
 
+type InputStates = "success" | "warning" | "error";
 type InputTypes = "small" | "normal" | "large";
+
+export const stateClasses = {
+  success: 'has-success',
+  warning: 'has-warning',
+  error: 'has-error',
+};
 
 type SelectOption = {
   label: string,
@@ -18,6 +25,7 @@ type Props = {
   label?: string,
   name?: string,
   inline?: boolean,
+  state?: InputStates,
   className?: string,
   style?: object,
   required?: boolean,
@@ -26,6 +34,7 @@ type Props = {
   options?: Array<SelectOption>,
   onChange?: (e: object) => void,
   value?: string,
+  disabled?: boolean,
   type?: InputTypes,
   /** Qa id */
   qa?: string,
@@ -49,8 +58,10 @@ class Select extends Component<Props> {
       name,
       inline,
       className,
+      state,
       style,
       required,
+      disabled,
       placeholder,
       options,
       onChange,
@@ -65,6 +76,7 @@ class Select extends Component<Props> {
       'has-icon-right',
       {
         'a-input--inline': !!inline,
+        [`${stateClasses[state]}`]: !!state,
         'is-required': !!required,
         [`a-input--${type}`]: !!type,
       },
@@ -80,7 +92,7 @@ class Select extends Component<Props> {
       <div className={selectClass} name={name} data-qa={qa}>
         {label && <InputLabel inline={inline} htmlFor={id}>{label}</InputLabel>}
         <div className={wrapperClass}>
-          <select id={id} name={name} onChange={(e) => this.onChange(e)} value={value}>
+          <select id={id} name={name} onChange={(e) => this.onChange(e)} value={value} disabled={disabled}>
             {placeholder && <option value="" selected disabled>{placeholder}</option>}
             {options.map(option => (
               <Option