+++ CHANGELOG.md
@@ -5,9 +5,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added select disabled state
+- `Select` add type
+
+### Fixed
+- `Textarea` Fixed an issue where required property was not being passed properly
+- `AccordeonTab` Added an option to pass a custom toggle function so that state can be managed in parent components.
 
 ### Fixed
 - `Select` Fix select elements must be either controlled or uncontrolled react warning
+- `Autocomplete` component now can handle default values when using async loading items
+### Added
+- `Autocomplete` supports placeholder and required properties
 
 ## v4.5.0 - 03-09-2020